refactor(moderation): add factory methods to enums and classifiers
- All enums: add random() for testing/seeding
- Severity: add weight() method, removing duplicate severityWeight() from classifiers
- CaseStatus: add isOpen() helper
- AppealStatus: add isResolved() helper
- RuleBasedClassifier, OpenAiClassifier: add make() static factory
- Classifiers now use Severity::weight() instead of private method
