+++ app-modules/moderation/src/Enums/CaseStatus.php
@@ -11,4 +11,16 @@ enum CaseStatus: string
     case Resolved = 'resolved';
     case Escalated = 'escalated';
     case Dismissed = 'dismissed';
+
+    public static function random(): self
+    {
+        $cases = self::cases();
+
+        return $cases[array_rand($cases)];
+    }
+
+    public function isOpen(): bool
+    {
+        return in_array($this, [self::Pending, self::Assigned]);
+    }
 }
