+++ app-modules/moderation/src/Enums/Platform.php
@@ -11,4 +11,11 @@ enum Platform: string
     case GitHub = 'github';
     case Twitter = 'twitter';
     case Web = 'web';
+
+    public static function random(): self
+    {
+        $cases = self::cases();
+
+        return $cases[array_rand($cases)];
+    }
 }
