+++ app-modules/moderation/tests/Feature/CaseWorkflowTest.php
@@ -3,9 +3,9 @@
 declare(strict_types=1);
 
 use He4rt\Identity\User\Models\User;
+use He4rt\Moderation\Cases\Actions\SubmitReport;
 use He4rt\Moderation\Cases\Models\ModerationCase;
 use He4rt\Moderation\Cases\Models\ModerationReport;
-use He4rt\Moderation\Cases\SubmitReport;
 use He4rt\Moderation\DTOs\ModerationContentDTO;
 use He4rt\Moderation\Enums\CaseSource;
 use He4rt\Moderation\Enums\Platform;
